Drop the -MG option from .d file generation
Since we are using -MT to make both the .d file and the .beam file
depend on the source files, adding -MG (add missing include files as
dependencies even if they are missing, assuming they will be generated)
will cause Make to loop if an include file does not exist - e.g. due to
a typo or a moved file - because it will try to regenerate the .d file
when it finds that the header file is still missing.